FISH-10659 Use a different option in AsyncServerSocketOptionsTest
SO_REUSEPORT is not a supported option in Java 8.
The test itself is not trying to test setting that option, there is another test for that, the test is just trying to test setting a supported option.
SO_REUSEADDR is available in both Java versions, so I have set it to use that.

Signed-off-by: Andrew Pielage <pandrex247@hotmail.com>